+++ backend/src/controller/plantings.rs
@@ -85,7 +85,7 @@ pub async fn create(
         .broadcast(
             path.into_inner(),
             Action::CreatePlanting(CreatePlantActionPayload::new(
-                dto,
+                dto.clone(),
                 user_info.id,
                 new_planting.action_id,
             )),
@@ -126,19 +126,19 @@ pub async fn update(
 
     let action = match update_planting {
         UpdatePlantingDto::Transform(action_dto) => Action::TransformPlanting(
-            TransformPlantActionPayload::new(planting, user_info.id, action_dto.action_id),
+            TransformPlantActionPayload::new(&planting, user_info.id, action_dto.action_id),
         ),
         UpdatePlantingDto::Move(action_dto) => Action::MovePlanting(MovePlantActionPayload::new(
-            planting,
+            &planting,
             user_info.id,
             action_dto.action_id,
         )),
         UpdatePlantingDto::UpdateAddDate(action_dto) => Action::UpdatePlantingAddDate(
-            UpdatePlantingAddDateActionPayload::new(planting, user_info.id, action_dto.action_id),
+            UpdatePlantingAddDateActionPayload::new(&planting, user_info.id, action_dto.action_id),
         ),
         UpdatePlantingDto::UpdateRemoveDate(action_dto) => {
             Action::UpdatePlantingRemoveDate(UpdatePlantingRemoveDateActionPayload::new(
-                planting,
+                &planting,
                 user_info.id,
                 action_dto.action_id,
             ))
@@ -147,7 +147,7 @@ pub async fn update(
 
     app_data.broadcaster.broadcast(map_id, action).await;
 
-    Ok(HttpResponse::Ok().json(planting))
+    Ok(HttpResponse::Ok().json(planting.clone()))
 }
 
 /// Endpoint for deleting a `Planting`.