+++ backend/src/model/dto/actions.rs
@@ -39,6 +39,8 @@ pub enum Action {
     UpdatePlantingAddDate(UpdatePlantingAddDateActionPayload),
     /// An action used to update the `remove_date` of a plant.
     UpdatePlantingRemoveDate(UpdatePlantingRemoveDateActionPayload),
+    /// An action used to update the `additional_name` of a plant.
+    UpdatePlantingAdditionalName(UpdatePlantingAdditionalNamePayload),
 }
 
 impl Action {
@@ -55,6 +57,7 @@ impl Action {
             Self::DeleteBaseLayerImage(payload) => payload.action_id,
             Self::UpdatePlantingAddDate(payload) => payload.action_id,
             Self::UpdatePlantingRemoveDate(payload) => payload.action_id,
+            Self::UpdatePlantingAdditionalName(payload) => payload.action_id,
         }
     }
 }
@@ -80,6 +83,7 @@ pub struct CreatePlantActionPayload {
     add_date: Option<NaiveDate>,
     remove_date: Option<NaiveDate>,
     seed_id: Option<i32>,
+    additional_name: Option<String>,
 }
 
 impl CreatePlantActionPayload {
@@ -101,6 +105,7 @@ impl CreatePlantActionPayload {
             add_date: payload.add_date,
             remove_date: payload.remove_date,
             seed_id: payload.seed_id,
+            additional_name: payload.additional_name,
         }
     }
 }
@@ -140,7 +145,7 @@ pub struct MovePlantActionPayload {
 
 impl MovePlantActionPayload {
     #[must_use]
-    pub fn new(payload: PlantingDto, user_id: Uuid, action_id: Uuid) -> Self {
+    pub fn new(payload: &PlantingDto, user_id: Uuid, action_id: Uuid) -> Self {
         Self {
             user_id,
             action_id,
@@ -168,7 +173,7 @@ pub struct TransformPlantActionPayload {
 
 impl TransformPlantActionPayload {
     #[must_use]
-    pub fn new(payload: PlantingDto, user_id: Uuid, action_id: Uuid) -> Self {
+    pub fn new(payload: &PlantingDto, user_id: Uuid, action_id: Uuid) -> Self {
         Self {
             user_id,
             action_id,
@@ -275,7 +280,7 @@ pub struct UpdatePlantingAddDateActionPayload {
 
 impl UpdatePlantingAddDateActionPayload {
     #[must_use]
-    pub fn new(payload: PlantingDto, user_id: Uuid, action_id: Uuid) -> Self {
+    pub fn new(payload: &PlantingDto, user_id: Uuid, action_id: Uuid) -> Self {
         Self {
             user_id,
             action_id,
@@ -298,12 +303,40 @@ pub struct UpdatePlantingRemoveDateActionPayload {
 
 impl UpdatePlantingRemoveDateActionPayload {
     #[must_use]
-    pub fn new(payload: PlantingDto, user_id: Uuid, action_id: Uuid) -> Self {
+    pub fn new(payload: &PlantingDto, user_id: Uuid, action_id: Uuid) -> Self {
         Self {
             user_id,
             action_id,
             id: payload.id,
             remove_date: payload.remove_date,
+        }
+    }
+}
+
+#[typeshare]
+#[derive(Debug, Serialize, Clone)]
+/// The payload of the [`Action::UpdatePlantingRemoveDate`].
+#[serde(rename_all = "camelCase")]
+pub struct UpdatePlantingAdditionalNamePayload {
+    user_id: Uuid,
+    action_id: Uuid,
+    id: Uuid,
+    additional_name: Option<String>,
+}
+
+impl UpdatePlantingAdditionalNamePayload {
+    #[must_use]
+    pub fn new(
+        payload: &PlantingDto,
+        new_additional_name: Option<String>,
+        user_id: Uuid,
+        action_id: Uuid,
+    ) -> Self {
+        Self {
+            user_id,
+            action_id,
+            id: payload.id,
+            additional_name: new_additional_name,
         }
     }
 }